Command
Unsure what to do since all we want to do for accelerometer is to measure acceleration, and since the subsystem already displays values, there's not much else we want to do in command. Also need to double check if periodic needs to be called in the first place or not.